+++ cli/cli/commands/test/testing_machinery/test_execution/test_executor/test_executor.go
@@ -183,11 +183,11 @@ func RunTest(
 		testSetupExecutionCtx,
 		log,
 		dockerManager,
+		enclaveId,
 		networkId,
 		testsuiteContainerName,
 		kurtosisApiIp,
 		testsuiteIpAddr,
-		enclaveId,
 		enclaveObjLabelsProvider.ForTestRunningTestsuiteContainer(),
 	)
 	if err != nil {